==> R Code/99_run_analysis.R <==
## -----------------------------------------------------------------------------
## Title: Run the Entire Analysis
##
## Author: Gen-Chang Hsu
##
## Date: 2022-05-31
##
## Description: Source and run all the R scripts in a logical order for complete 
##              analysis results
## 
## -----------------------------------------------------------------------------
source("./R Code/01_data_preparation.R")
source("./R Code/02_SEAb.R")
source("./R Code/03_figures.R")
source("./R Code/04_PERMANOVA_PERMADISP.R")
